server/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.1
email-validator==2.1.0.post1
marshmallow==3.20.1
bcrypt==4.1.2
requests==2.31.0
openai==1.30.1  # if AI budgeting, sentiment, fraud detection features use OpenAI

 